Whitespace corrections
While this is a big ugly commit that I would rather not do, I am
forced as it's completely impossible to add functionality in a clean
way without messing up other lines.

So, all files are converted to UNIX line endings (some were in DOS)
and all files have been formatted in Eclipse using the built-in
formatter.